requirements.txt
certifi==2017.4.17
chardet==3.0.4
cycler==0.10.0
definitions==0.2.0
enum34==1.1.6
idna==2.5
lxml==3.8.0
matplotlib==2.0.2
nltk==3.2.4
numpy==1.13.0
olefile==0.44
pandas==0.20.2
Pillow==4.1.1
prettytensor==0.7.4
protobuf==3.3.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
requests==2.17.3
scikit-learn==0.18.2
scipy==0.19.1
sets==0.3.2
six==1.10.0
sklearn==0.0
tensorflow==1.1.0
urllib3==1.21.1
Werkzeug==0.12.2
